Split transform logic into separate crate
