+++ swc/swc-plugin-css-variable/Cargo.toml
@@ -5,23 +5,13 @@ edition = "2021"
 
 [lib]
 # cdylib is for dynamic system library and necessary to produce the Wasm plugin
-# rlib is required for criterion benchmark
-crate-type = ["cdylib", "rlib"]
+crate-type = ["cdylib"]
 
 [dependencies]
-base62 = "2.0.2"
 lazy_static = "1.4.0"
 pathdiff = "0.2.1"
 regex = "1.7.1"
-serde = { version = "1.0.152", features = ["derive"] }
 serde_json = "1.0.91"
 swc_core = { version = "0.58.0", features = ["ecma_plugin_transform"] }
-xxhash-rust = { version = "0.8.6", features = ["xxh32"] }
 
-[dev-dependencies]
-criterion = { version = "0.4.0", default-features = false }
-swc_core = { version = "0.58.0", features = ["ecma_parser"] }
-
-[[bench]]
-name = "bench_main"
-harness = false
+transform = { path = "../transform" }
