+++ swc/transform/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "transform"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+base62 = "2.0.2"
+serde = { version = "1.0.152", features = ["derive"] }
+swc_core = { version = "0.58.0", features = ["ecma_ast"] }
+xxhash-rust = { version = "0.8.6", features = ["xxh32"] }
+
+[dev-dependencies]
+criterion = { version = "0.4.0", default-features = false }
+swc_core = { version = "0.58.0", features = ["ecma_parser"] }
+
+[[bench]]
+name = "bench_main"
+harness = false
