+++ swc/transform/benches/bench_main.rs
@@ -2,14 +2,14 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 
 pub fn short_hash(c: &mut Criterion) {
     c.bench_function("short filename hash", |b| {
-        b.iter(|| swc_plugin_css_variable::hash(black_box("fileName.ts")))
+        b.iter(|| transform::hash(black_box("fileName.ts")))
     });
 }
 
 pub fn longer_hash(c: &mut Criterion) {
     c.bench_function("longer filename hash", |b| {
         b.iter(|| {
-            swc_plugin_css_variable::hash(
+            transform::hash(
                 black_box("some/slightly/somewhat/much/very/much/longer/harder/better/faster/stronger/convoluted/fileName.ts")
             )
         })
