+++ swc/transform/src/hash.rs
